Issue #SB-15614 fix: Telemetry events router removing duplicate code
